# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all published blog posts
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag
        - name: search
          in: query
          schema:
            type: string
          description: Search by title or content
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get a blog post by ID
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Update a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Delete a blog post
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /interactions/like:
    post:
      operationId: toggleLike
      tags: [interactions]
      summary: Toggle like on a project or post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionInput"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /interactions/save:
    post:
      operationId: toggleSave
      tags: [interactions]
      summary: Toggle save/bookmark on a project or post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionInput"
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /interactions/saved:
    get:
      operationId: getSaved
      tags: [interactions]
      summary: Get all items saved by the current user
      responses:
        "200":
          description: Saved items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedItemsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [email, password, username]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          minLength: 3
        displayName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, username, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        githubUsername:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    Project:
      type: object
      required: [id, userId, title, createdAt, likeCount, saveCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        liveUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        likeCount:
          type: integer
        saveCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Post:
      type: object
      required: [id, userId, title, published, createdAt, likeCount, saveCount]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        category:
          type: ["string", "null"]
        published:
          type: boolean
        likeCount:
          type: integer
        saveCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        isSaved:
          type: ["boolean", "null"]
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostListResponse:
      type: object
      required: [posts, total, page, limit]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PostInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        published:
          type: boolean
          default: false
    PostUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        published:
          type: boolean
    InteractionInput:
      type: object
      required: [targetType, targetId]
      properties:
        targetType:
          type: string
          enum: [project, post]
        targetId:
          type: integer
    InteractionResponse:
      type: object
      required: [active, count]
      properties:
        active:
          type: boolean
        count:
          type: integer
    SavedItem:
      type: object
      required: [id, targetType, targetId, createdAt]
      properties:
        id:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        project:
          $ref: "#/components/schemas/Project"
        post:
          $ref: "#/components/schemas/Post"
        createdAt:
          type: string
          format: date-time
    SavedItemsResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SavedItem"
...